refactor(core.py, utils.py): replace wait_for_resources* with wait_for_ready
