+++ src/hassette/utils.py
@@ -4,13 +4,45 @@ import typing
 from logging import getLogger
 
 import aiohttp
-import anyio
 from whenever import OffsetDateTime, SystemDateTime, ZonedDateTime
 
 if typing.TYPE_CHECKING:
     from hassette.core.classes import Resource
 
 LOGGER = getLogger(__name__)
+
+
+async def wait_for_ready(
+    resources: "list[Resource] | Resource",
+    poll_interval: float = 0.1,
+    timeout: int = 20,
+    shutdown_event: asyncio.Event | None = None,
+) -> bool:
+    """Block until all dependent resources are ready or shutdown is requested.
+
+    Args:
+        resources (list[Resource] | Resource): The resources to wait for.
+        poll_interval (float): The interval to poll for resource status.
+        timeout (int): The timeout for the wait operation.
+
+    Returns:
+        bool: True if all resources are ready, False if timeout or shutdown.
+
+    Raises:
+        CancelledError: If the wait operation is cancelled.
+        TimeoutError: If the wait operation times out.
+    """
+
+    resources = resources if isinstance(resources, list) else [resources]
+    deadline = asyncio.get_event_loop().time() + timeout
+    while True:
+        if shutdown_event and shutdown_event.is_set():
+            return False
+        if all(r.is_ready() for r in resources):
+            return True
+        if asyncio.get_event_loop().time() >= deadline:
+            return False
+        await asyncio.sleep(poll_interval)
 
 
 def convert_utc_timestamp_to_system_tz(timestamp: int | float) -> SystemDateTime:
@@ -37,104 +69,6 @@ def convert_datetime_str_to_system_tz(value: str | SystemDateTime | None) -> Sys
     if value is None or isinstance(value, SystemDateTime):
         return value
     return OffsetDateTime.parse_common_iso(value).to_system_tz()
-
-
-async def wait_for_resources_running_or_raise(
-    resources: list["Resource"],
-    poll_interval: float = 0.1,
-    timeout: int = 20,
-    shutdown_event: asyncio.Event | None = None,
-) -> None:
-    """Block until all dependent resources are running or shutdown is requested.
-
-    Args:
-        resources (list[Resource]): The resources to wait for.
-        poll_interval (float): The interval to poll for resource status.
-        timeout (int): The timeout for the wait operation.
-        shutdown_event (asyncio.Event | None): Optional event to signal shutdown.
-
-    Raises:
-        RuntimeError: If any resource fails to start or timeout occurs.
-    """
-    from hassette.core.enums import ResourceStatus
-
-    results = await wait_for_resources_running(
-        resources, poll_interval=poll_interval, timeout=timeout, shutdown_event=shutdown_event
-    )
-    if not results:
-        failed_to_start = [r.class_name for r in resources if r.status != ResourceStatus.RUNNING]
-        LOGGER.error("One or more resources failed to start: %s", ", ".join(failed_to_start))
-        raise RuntimeError(f"One or more resources failed to start: {', '.join(failed_to_start)}")
-
-
-async def wait_for_resources_running(
-    resources: list["Resource"],
-    poll_interval: float = 0.1,
-    timeout: int = 20,
-    shutdown_event: asyncio.Event | None = None,
-) -> bool:
-    """Block until all dependent resources are running or shutdown is requested.
-
-    Args:
-        resources (list[Resource]): The resources to wait for.
-        poll_interval (float): The interval to poll for resource status.
-        timeout (int): The timeout for the wait operation.
-        shutdown_event (asyncio.Event | None): Optional event to signal shutdown.
-
-    Returns:
-        bool: True if all resources are running, False if shutdown is requested.
-    """
-    futures = [
-        wait_for_resource_running(resource, poll_interval=poll_interval, timeout=timeout, shutdown_event=shutdown_event)
-        for resource in resources
-    ]
-
-    try:
-        results = await asyncio.gather(*futures, return_exceptions=True)
-    except Exception as e:
-        LOGGER.error("Error waiting for resources: %s", e)
-        return False
-
-    # Convert exceptions to False, log errors
-    for i, result in enumerate(results):
-        if isinstance(result, Exception):
-            LOGGER.error("Error waiting for resource: %s", result)
-            results[i] = False
-    return all(results)
-
-
-async def wait_for_resource_running(
-    resource: "Resource", poll_interval: float = 0.1, timeout: int = 20, shutdown_event: asyncio.Event | None = None
-) -> bool:
-    """Block until a dependent resource is running or shutdown is requested.
-
-    Args:
-        resource (Resource): The resource to wait for.
-        poll_interval (float): The interval to poll for resource status.
-        timeout (int): The timeout for the wait operation.
-        shutdown_event (asyncio.Event | None): Optional event to signal shutdown.
-
-    Returns:
-        bool: True if the resource is running, False if shutdown.
-    """
-    from hassette.core.enums import ResourceStatus
-
-    with anyio.move_on_after(timeout) as cancel_scope:
-        while resource.status != ResourceStatus.RUNNING:
-            if shutdown_event and shutdown_event.is_set():
-                LOGGER.warning("Shutdown in progress, aborting wait for resource '%s'", resource.class_name)
-                return False
-            await asyncio.sleep(poll_interval)
-
-    if cancel_scope.cancel_called:
-        LOGGER.error("Timeout waiting for resource '%s' to start after %d seconds", resource.class_name, timeout)
-        return False
-
-    if resource.status != ResourceStatus.RUNNING:
-        LOGGER.error("Resource '%s' is not running", resource.class_name)
-        return False
-
-    return True
 
 
 def get_traceback_string(exception: Exception) -> str:
